+++ components/faphub/dao/network/src/main/java/com/flipperdevices/faphub/dao/network/ktorfit/model/KtorfitReport.kt
@@ -1,0 +1,10 @@
+package com.flipperdevices.faphub.dao.network.ktorfit.model
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class KtorfitReport(
+    @SerialName("description") val description: String,
+    @SerialName("description_type") val descriptionType: String
+)