+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - [Feature] Now the installed apps are even shown offline
 - [Feature] Change design for installed tab
 - [Feature] Add metric for firmware origin
+- [Feature] Add more error catching for FapHub
 - [Refactor] Migrate all feature components modules from KSP to Anvil
 - [Refactor] Migrate to Ktorfit
 - [Refactor] Migrate key emulate to new module, decompose Emulate Helper